CM_HOUSE_TELEPORT: (byte) cast house.GetTeleportHeading() to TeleportService.TeleportTo byte heading param (byte<->sbyte vein; Java byte heading=sbyte); clears 2 CS1503 (743->741)

+++ dotnetConversion/src/Aion.GameServer/Network/Aion/ClientPackets/CM_HOUSE_TELEPORT.cs
@@ -89,7 +89,7 @@ public class CM_HOUSE_TELEPORT : AionClientPacket
             return;
 
         WorldMapInstance instance = InstanceService.GetOrCreateHouseInstance(house);
-        TeleportService.TeleportTo(player, instance, house.GetX(), house.GetY(), house.GetZ(), house.GetTeleportHeading(),
+        TeleportService.TeleportTo(player, instance, house.GetX(), house.GetY(), house.GetZ(), (byte)house.GetTeleportHeading(),
             TeleportAnimation.FADE_OUT_BEAM);
     }
